Data Vizualization and Analyses Update
Finished vizualization and testing of certain statistical techniques. Testing and t.test show that imposters and players show a significant difference at  6 players left in the game. This indicates a way of scoring players could be developed. Decided on z-scores as conservative estimates and population simulation calculations for liberal estimates.